fix size_t

+++ dipu/torch_dipu/csrc_dipu/runtime/core/DIPUOpInferrer.cpp
@@ -252,9 +252,6 @@ void OpInferrer::compute_perm() {
         // move on to the next input if one of the dimensions is broadcasted
         continue;
       }
-      //       if (stride0 != stride1) {
-      //         return stride0 > stride1 ? 1 : -1;
-      //       }
       if (stride0 < stride1) {
         return -1;
       }
@@ -271,8 +268,8 @@ void OpInferrer::compute_perm() {
 
   // insertion sort with support for ambiguous comparisons
   for (const auto i : c10::irange(ndim())) {
-    int dim1 = i;
-    for (int dim0 = i - 1; dim0 >= 0; dim0--) {
+    size_t dim1 = i;
+    for (size_t dim0 = i - 1; dim0 >= 0; dim0--) {
       int comparison = should_swap(perm_[dim0], perm_[dim1]);
       if (comparison > 0) {
         std::swap(perm_[dim0], perm_[dim1]);
